open current page in the publish dialog
